feat: refactor password reset functionality by consolidating email service integration, updating profile controller and repository, and enhancing API documentation for password reset endpoints
