+++ src/auth/auth.controller.ts
@@ -1,6 +1,7 @@
 import { Body, Controller, Post } from "@nestjs/common";
 import { ApiBody, ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
 
+import EmailService from "@/common/services/email.service";
 import { AuthData, ResponseCode, ResponseWrapper, TokensData } from "@/types";
 import { buildResponse } from "@/utils/response/response-wrapper";
 
@@ -18,10 +19,7 @@ import {
   AuthCredentialsDto,
   AuthSignInDto,
   AuthSocialDto,
-  ForgotPasswordDto,
-  ResetPasswordDto,
 } from "./dto/auth-credentials.dto";
-import EmailService from "@/common/services/email.service";
 
 @ApiTags("Authentication")
 @Controller("auth")
@@ -86,86 +84,5 @@ export class AuthController {
   ): Promise<ResponseWrapper<Pick<TokensData, "access_token" | "access_token_expires">>> {
     const data = await this.authService.refreshAccessToken(refreshToken);
     return buildResponse({data, code:ResponseCode.OK, message:"Access token refreshed"});
-  }
-
-  @Post('/forgot-password')
-  @ApiOperation({ summary: "Request password reset email" })
-  @ApiBody({
-    schema: {
-      type: 'object',
-      properties: {
-        email: {
-          type: 'string',
-          format: 'email',
-          description: 'User email address',
-          example: 'user@example.com'
-        }
-      },
-      required: ['email']
-    }
-  })
-  @ApiResponse({
-    status: 200,
-    description: "Password reset email sent",
-    schema: {
-      type: 'object',
-      properties: {
-        success: { type: 'boolean' },
-        message: { type: 'string' },
-        data: { type: 'object' },
-        code: { type: 'string' }
-      }
-    }
-  })
-  @ApiResponse({ status: 400, description: "Bad request - invalid email" })
-  async forgotPassword(
-    @Body() forgotPasswordDto: ForgotPasswordDto,
-  ): Promise<ResponseWrapper<{ success: boolean; message: string }>> {
-    const data = await this.emailService.sendForgotPasswordEmail(forgotPasswordDto.email);
-    return buildResponse({ data, code: ResponseCode.OK, message: data.message });
-  }
-
-  @Post('/reset-password')
-  @ApiOperation({ summary: "Reset password using token" })
-  @ApiBody({
-    schema: {
-      type: 'object',
-      properties: {
-        token: {
-          type: 'string',
-          description: 'Password reset token from email',
-          example: 'abc123def456...'
-        },
-        newPassword: {
-          type: 'string',
-          description: 'New password (must be strong)',
-          example: 'NewStrongPass123!',
-          minLength: 8,
-          maxLength: 32
-        }
-      },
-      required: ['token', 'newPassword']
-    }
-  })
-  @ApiResponse({
-    status: 200,
-    description: "Password reset successfully",
-    schema: {
-      type: 'object',
-      properties: {
-        success: { type: 'boolean' },
-        message: { type: 'string' },
-        data: { type: 'object' },
-        code: { type: 'string' }
-      }
-    }
-  })
-  @ApiResponse({ status: 400, description: "Bad request - invalid data" })
-  @ApiResponse({ status: 404, description: "Not found - invalid or expired token" })
-  async resetPassword(
-    @Body() resetPasswordDto: ResetPasswordDto,
-  ): Promise<ResponseWrapper<{ success: boolean; message: string }>> {
-    const data = await this.emailService.resetPassword(resetPasswordDto.token, resetPasswordDto.newPassword);
-    return buildResponse({ data, code: ResponseCode.OK, message: data.message });
   }
 }
