--- src/common/services/forgot-password-email.html
@@ -1,179 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Reset Your Password - Cinematik</title>
-    <style>
-        * {
-            margin: 0;
-            padding: 0;
-            box-sizing: border-box;
-        }
-        body {
-            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, sans-serif;
-            line-height: 1.6;
-            color: #333;
-            background-color: #f8f9fa;
-        }
-        .container {
-            max-width: 600px;
-            margin: 0 auto;
-            background-color: #ffffff;
-            border-radius: 12px;
-            overflow: hidden;
-            box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-        }
-        .header {
-            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-            padding: 40px 30px;
-            text-align: center;
-        }
-        .header h1 {
-            color: #ffffff;
-            font-size: 28px;
-            font-weight: 700;
-            margin-bottom: 10px;
-        }
-        .header p {
-            color: rgba(255, 255, 255, 0.9);
-            font-size: 16px;
-        }
-        .content {
-            padding: 40px 30px;
-        }
-        .greeting {
-            font-size: 18px;
-            font-weight: 600;
-            margin-bottom: 20px;
-            color: #2d3748;
-        }
-        .message {
-            font-size: 16px;
-            color: #4a5568;
-            margin-bottom: 30px;
-            line-height: 1.7;
-        }
-        .reset-button {
-            display: inline-block;
-            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-            color: #ffffff;
-            text-decoration: none;
-            padding: 16px 32px;
-            border-radius: 8px;
-            font-weight: 600;
-            font-size: 16px;
-            text-align: center;
-            margin: 20px 0;
-            transition: transform 0.2s ease;
-        }
-        .reset-button:hover {
-            transform: translateY(-2px);
-        }
-        .warning {
-            background-color: #fff5f5;
-            border: 1px solid #fed7d7;
-            border-radius: 8px;
-            padding: 20px;
-            margin: 30px 0;
-        }
-        .warning h3 {
-            color: #c53030;
-            font-size: 16px;
-            margin-bottom: 10px;
-        }
-        .warning p {
-            color: #742a2a;
-            font-size: 14px;
-            margin: 0;
-        }
-        .footer {
-            background-color: #f7fafc;
-            padding: 30px;
-            text-align: center;
-            border-top: 1px solid #e2e8f0;
-        }
-        .footer p {
-            color: #718096;
-            font-size: 14px;
-            margin-bottom: 10px;
-        }
-        .footer a {
-            color: #667eea;
-            text-decoration: none;
-        }
-        .footer a:hover {
-            text-decoration: underline;
-        }
-        .expiry-note {
-            background-color: #ebf8ff;
-            border: 1px solid #bee3f8;
-            border-radius: 8px;
-            padding: 15px;
-            margin: 20px 0;
-        }
-        .expiry-note p {
-            color: #2b6cb0;
-            font-size: 14px;
-            margin: 0;
-        }
-        @media (max-width: 600px) {
-            .container {
-                margin: 10px;
-                border-radius: 8px;
-            }
-            .header, .content, .footer {
-                padding: 20px;
-            }
-            .header h1 {
-                font-size: 24px;
-            }
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <div class="header">
-            <h1>🎬 Cinematik</h1>
-            <p>Your Ultimate Movie Experience</p>
-        </div>
-        <div class="content">
-            <div class="greeting">Hello {{userName}},</div>
-            <div class="message">
-                We received a request to reset your password for your Cinematik account. 
-                If you didn't make this request, you can safely ignore this email.
-            </div>
-            <div style="text-align: center;">
-                <a href="{{resetLink}}" class="reset-button">
-                    Reset Your Password
-                </a>
-            </div>
-            <div class="expiry-note">
-                <p><strong>⏰ This link will expire in 1 hour for security reasons.</strong></p>
-            </div>
-            <div class="warning">
-                <h3>🔒 Security Notice</h3>
-                <p>
-                    If you didn't request this password reset, please contact our support team immediately. 
-                    Your account security is important to us.
-                </p>
-            </div>
-            <div class="message">
-                If the button above doesn't work, you can copy and paste the following link into your browser:
-                <br><br>
-                <a href="{{resetLink}}" style="color: #667eea; word-break: break-all;">{{resetLink}}</a>
-            </div>
-        </div>
-        <div class="footer">
-            <p>Best regards,<br>The Cinematik Team</p>
-            <p>
-                <a href="mailto:support@cinematik.com">support@cinematik.com</a> | 
-                <a href="https://cinematik.com">cinematik.com</a>
-            </p>
-            <p style="font-size: 12px; color: #a0aec0;">
-                This email was sent to you because you requested a password reset for your Cinematik account.
-            </p>
-        </div>
-    </div>
-</body>
-</html>
