+++ src/common/templates/email-styles.hbs
@@ -1,0 +1,27 @@
+<style>
+  * { margin: 0; padding: 0; box-sizing: border-box; }
+  body { font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, sans-serif; line-height: 1.6; color: #333; background-color: #f8f9fa; }
+  .container { max-width: 600px; margin: 0 auto; background-color: #ffffff; border-radius: 12px; overflow: hidden; box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); }
+  .header { background: linear-gradient(135deg, #667eea 0%, #764ba2 100%); padding: 40px 30px; text-align: center; }
+  .header h1 { color: #ffffff; font-size: 28px; font-weight: 700; margin-bottom: 10px; }
+  .header p { color: rgba(255, 255, 255, 0.9); font-size: 16px; }
+  .content { padding: 40px 30px; }
+  .greeting { font-size: 18px; font-weight: 600; margin-bottom: 20px; color: #2d3748; }
+  .message { font-size: 16px; color: #4a5568; margin-bottom: 30px; line-height: 1.7; }
+  .reset-button { display: inline-block; background: linear-gradient(135deg, #667eea 0%, #764ba2 100%); color: #ffffff; text-decoration: none; padding: 16px 32px; border-radius: 8px; font-weight: 600; font-size: 16px; text-align: center; margin: 20px 0; transition: transform 0.2s ease; }
+  .reset-button:hover { transform: translateY(-2px); }
+  .warning { background-color: #fff5f5; border: 1px solid #fed7d7; border-radius: 8px; padding: 20px; margin: 30px 0; }
+  .warning h3 { color: #c53030; font-size: 16px; margin-bottom: 10px; }
+  .warning p { color: #742a2a; font-size: 14px; margin: 0; }
+  .footer { background-color: #f7fafc; padding: 30px; text-align: center; border-top: 1px solid #e2e8f0; }
+  .footer p { color: #718096; font-size: 14px; margin-bottom: 10px; }
+  .footer a { color: #667eea; text-decoration: none; }
+  .footer a:hover { text-decoration: underline; }
+  .expiry-note { background-color: #ebf8ff; border: 1px solid #bee3f8; border-radius: 8px; padding: 15px; margin: 20px 0; }
+  .expiry-note p { color: #2b6cb0; font-size: 14px; margin: 0; }
+  @media (max-width: 600px) {
+    .container { margin: 10px; border-radius: 8px; }
+    .header, .content, .footer { padding: 20px; }
+    .header h1 { font-size: 24px; }
+  }
+</style>
