+++ src/common/templates/forgot-password-email.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Reset Your Password - Cinematik</title>
+   {{> styles }}
+</head>
+<body>
+    <div class="container">
+        <div class="header">
+            <h1>🎬 Cinematik</h1>
+            <p>Your Ultimate Movie Experience</p>
+        </div>
+        <div class="content">
+            <div class="greeting">Hello {{userName}},</div>
+            <div class="message">
+                We received a request to reset your password for your Cinematik account. 
+                If you didn't make this request, you can safely ignore this email.
+            </div>
+            <div style="text-align: center;">
+                <a href="{{resetLink}}" class="reset-button">
+                    Reset Your Password
+                </a>
+            </div>
+            <div class="expiry-note">
+                <p><strong>⏰ This link will expire in 1 hour for security reasons.</strong></p>
+            </div>
+            <div class="warning">
+                <h3>🔒 Security Notice</h3>
+                <p>
+                    If you didn't request this password reset, please contact our support team immediately. 
+                    Your account security is important to us.
+                </p>
+            </div>
+            <div class="message">
+                If the button above doesn't work, you can copy and paste the following link into your browser:
+                <br><br>
+                <a href="{{resetLink}}" style="color: #667eea; word-break: break-all;">{{resetLink}}</a>
+            </div>
+        </div>
+        <div class="footer">
+            <p>Best regards,<br>The Cinematik Team</p>
+            <p>
+                <a href="mailto:support@cinematik.com">support@cinematik.com</a> | 
+                <a href="https://cinematik.com">cinematik.com</a>
+            </p>
+            <p style="font-size: 12px; color: #a0aec0;">
+                This email was sent to you because you requested a password reset for your Cinematik account.
+            </p>
+        </div>
+    </div>
+</body>
+</html>
