--- src/profile/dto/update-profile.dto.ts
@@ -1,11 +1,0 @@
-import { IsEmail, IsName } from "@/common/decorators/dto.decorators";
-
-class UpdateProfileDto {
-  @IsName()
-  name: string;
-
-  @IsEmail()
-  email: string;
-}
-
-export default UpdateProfileDto;
